Resolve Compiler Warnings (#64)

+++ Source/Core/Renderer/ERS_CLASS_VisualRenderer/ERS_CLASS_LightIconRenderer.cpp
@@ -139,7 +139,7 @@ void ERS_CLASS_LightIconRenderer::Draw(ERS_STRUCT_Camera* Camera, ERS_CLASS_Scen
 
 
     // Draw All Spot Lights
-    for (int i = 0; i < SceneManager->Scenes_[SceneManager->ActiveScene_]->SpotLights.size(); i++) {
+    for (int i = 0; (long)i < (long)SceneManager->Scenes_[SceneManager->ActiveScene_]->SpotLights.size(); i++) {
 
         glm::vec3 LightPosition = SceneManager->Scenes_[SceneManager->ActiveScene_]->SpotLights[i]->Pos;
         glm::mat4 NewModelMatrix = glm::translate(LightIconRendererModelArray_, LightPosition);
